+++ src/components/index.ts
@@ -6,3 +6,4 @@ export { default as SettingsDialog } from "./SettingsDialog";
 export { default as Footer } from "./Footer";
 export { default as Disclaimer } from "./Disclaimer";
 export { default as LlmDebugDialog } from "./LlmDebugDialog";
+export { default as ModelSelectionDialog } from "./ModelSelectionDialog";